disamb: fix non-determinism in create_dfa
Petgraph allows multiple edges between two nodes, but apparently not
multiple edges with the same edge value that share a node.

Edges were being shared between mergeable runs, which eventually
means that merge([ibid], [ibid, locator]) would sometimes give

*--ibid--$--loc---$

(correctly accepting [ibid] and [ibid, loc])

but sometimes give

*--ibid--*--loc---$

(not accepting [ibid])

depending on the iteration order of the edges or the FreeCondSets, each
of which is non-deterministic.

This is fixed by starting on unique nodes per run, so there is no common
starting node. When merged, the DFA is properly deterministic, so there
are no same-edge-value-different-targets.

For future reference, petgraph allows:

   a
*======*
   b

BUT NOT:

    a
 ,-----*
*
 `-----*
    a