fix nondeterminism in Dfa::accepts_data
The real problem all along (ref 32995f536) was that iterating
over NEIGHBOURS is not the same as iterating over EDGES.

If your neighbour is connected to you twice, then graph.neighbours(you)
will give you that neighbour once, and lookup_edge() will return whichever
edge petgraph's implementation decides to give you. (Depends on insertion
order). graph.edges(you), on the other hand, will give you both edges,
and each will have an EdgeRef::target() of the same neighbour.

There's now a test of this with two edges leading to the same neighbour,
which used to pass 50% of the time, now it's 100%.