Alloc year suffixes in bib order; and use ghost entries only, no ibid cites.
The bib order is per the spec, a long-standing TODO in year_suffixes().
Using ghost cites for every single reference is a new tactic, to avoid
the following case where @a and @b need a year suffix but @c is different:

uncited: @b, @c
---
@a
@a
@a

Here, the second @a is ibid, and only renders "ibid.". However, this cite
will be accepted by all three DFAs. So previously, the bibliography would
have a year suffix added to all three entries, simply because an ibid cite
in the document, not expected to be able to be disambiguated, matched each.

New code only using ghost cites (which are always First) means that these
suffixes are based on the most ideal representation of the reference.