Separate matching_refs from IrGen
Basically, IrGen contained matching_refs. Whenever you changed the list
of cites in the document or any individual reference, the matching_refs
had to be recomputed, because refs_accepting_cite needs
disamb_participants() and by extension every ref_dfa().

The new way means ir_gen0 rarely gets recomputed unless its matching
ref was changed, or if the cite itself was edited. (Note that the cite
id changes when the cite itself is edited, which is an issue for garbage
collection and/or optimising the interned CiteId to be ONLY based on
indexes within a cluster ID's cites, and recover the actual data
elsewhere.