rewrite some nom code not to need nom::error::Error: Clone
in nom v6, the parser combinators return FnMuts. So they have to be able
to be called again, and fold_many1's initial/accumulator value therefore
has to be Clone to allow it to be cloned on each FnMut invocation.

nom::error::Error is not Clone, so the IResult that was in the accumulator
before was not able to be cloned. This just skirts around that issue.
But it would be better if nom::error::Error were indeed Clone.