Fix bad bug in set_cluster_order
This was caused while attempting to throw errors for non-existent cluster IDs.
The "list of all clusters that won't panic when you get their cites/notes/etc"
and "list of all clusters in the document, per set_cluster_order"
were treated as the same thing, so removing something from the flow
caused it to panic when you attempted to re-add it.

Those two concepts are now separate.

Second, there is a breaking change to the preview_citation_cluster API,
which now accepts a PreviewCluster instead of a list of cites.
This is so you can preview in a specific cluster mode.